split runPredefinedEffectFunctions to multiple files
